requirements.txt
python-uinput
evdev
pygame
python-dateutil
ruff
pywinusb